+++ Civi/ActionProvider/Action/MailingEvent/MailingEventGetContactId.php
@@ -1,0 +1,83 @@
+<?php
+/**
+ * Copyright (C) 2024  Jaap Jansma (jaap.jansma@civicoop.org)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+namespace Civi\ActionProvider\Action\MailingEvent;
+
+use Civi\ActionProvider\Action\AbstractAction;
+use Civi\ActionProvider\Parameter\ParameterBagInterface;
+use Civi\ActionProvider\Parameter\Specification;
+use Civi\ActionProvider\Parameter\SpecificationBag;
+use CRM_ActionProvider_ExtensionUtil as E;
+
+class MailingEventGetContactId extends AbstractAction {
+
+  /**
+   * Run the action
+   *
+   * @param ParameterBagInterface $parameters
+   *   The parameters to this action.
+   * @param ParameterBagInterface $output
+   *   The parameters this action can send back
+   *
+   * @return void
+   */
+  protected function doAction(ParameterBagInterface $parameters, ParameterBagInterface $output) {
+    $hash = $parameters->getParameter('hash');
+    $queue_id = $parameters->getParameter('event_queue_id');
+    $job_id = $parameters->getParameter('job_id');
+    $q = \CRM_Mailing_Event_BAO_Queue::verify($job_id, $queue_id, $hash);
+    if ($q) {
+      $output->setParameter('contact_id', $q->contact_id);
+    }
+  }
+
+  /**
+   * Returns the specification of the configuration options for the actual
+   * action.
+   *
+   * @return SpecificationBag
+   */
+  public function getConfigurationSpecification() {
+    return new SpecificationBag([]);
+  }
+
+  /**
+   * Returns the specification of the parameters of the actual action.
+   *
+   * @return SpecificationBag
+   */
+  public function getParameterSpecification() {
+    return new SpecificationBag([
+      new Specification('hash', 'String', E::ts('Hash'), TRUE),
+      new Specification('job_id', 'Integer', E::ts('Job Id'), TRUE),
+      new Specification('event_queue_id', 'Integer', E::ts('Event Queue ID'), TRUE),
+    ]);
+  }
+
+  /**
+   * Returns the specification of the output parameters of this action.
+   *
+   * This function could be overridden by child classes.
+   *
+   * @return SpecificationBag
+   */
+  public function getOutputSpecification() {
+    return new SpecificationBag([new Specification('contact_id', 'Integer', E::ts("Contact ID"))]);
+  }
+
+}